Move overview controller to existing annotation controller
(it is now called "index")
