+++ app/abilities/annotation_ability.rb
@@ -2,7 +2,7 @@ class AnnotationAbility
   include CanCan::Ability
 
   def initialize(user)
-    can [:edit, :update, :destroy], Annotation do |annotation|
+    can [:index, :edit, :update, :destroy], Annotation do |annotation|
       annotation.user == user
     end
